refactor(workflow): own worktree baselines and file-change projection
Node dispatch and execution provenance were reading as one responsibility in
executor.rs. They are not: a node runs to completion whether or not a baseline
could be taken, and every failure on the provenance side degrades to an empty
diff rather than to a failed node. Keeping that rule beside the dispatch that
does not depend on it is what made "no baseline" easy to misread as "an empty
worktree in which every file is new".

Moves baseline capture, its size cap, the file-change diff and baseline
persistence into workflow/run/worktree.rs, together with the five tests that
own those invariants. executor.rs drops below the 800-line gate, so its size
exception is removed rather than lowered.
